# docs/swagger.yaml
definitions:
  handler.LoginRequestUser:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - role
    - username
    type: object
...
info:
  contact: {}
paths:
  /admin/register:
    post:
      consumes:
      - application/json
      description: Register a new admin with username, password, and role
      parameters:
      - description: Register admin request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful registration message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new admin
      tags:
      - Admin
...
  /customer/login:
    post:
      consumes:
      - application/json
      description: Authenticate a customer with email and password and return a JWT
        token
      parameters:
      - description: Login customer request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequestUser'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token response
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a customer
      tags:
      - Customer
  /customer/register:
    post:
      consumes:
      - application/json
      description: Register a new customer with name, username, email, and password
      parameters:
      - description: Register customer request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful registration message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new customer
      tags:
      - Customer
...
swagger: "2.0"